jobradar/requirements.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==4.2.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.7.0
mssql-django==1.4.1
pyodbc==5.0.1
python-decouple==3.8
coverage==7.4.1
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
MouseInfo==0.1.3
nltk==3.9.1
numpy==2.2.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytweening==1.2.0
pytz==2025.2
django-cors-headers==4.7.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
setuptools==80.3.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.20
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.3.7
wordcloud==1.9.4
wsproto==1.2.0
